third_party/WebKit/LayoutTests/usb/usb.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/fake-devices.js"></script>
<script src="resources/usb-helpers.js"></script>
<script>
'use strict';

usb_test(usb => {
  usb.mockDeviceManager.addMockDevice(usb.fakeDevices[0]);

  return navigator.usb.getDevices().then(devices => {
    assert_equals(devices.length, 1);
    usb.assertDeviceInfoEquals(devices[0], usb.fakeDevices[0]);
  });
}, 'getDevices returns devices exposed by the DeviceManager service.');

usb_test(usb => {
  return navigator.usb.requestDevice({ filters: [] })
    .then(device => {
      assert_unreachable('requestDevice should reject without a user gesture');
    })
    .catch(error => {
      assert_equals(error.code, DOMException.SECURITY_ERR);
    });
}, 'requestDevice rejects when called without a user gesture');

usb_test(usb => {
  usb.mockChooserService.setChosenDevice(null);
  return callWithKeyDown(() => navigator.usb.requestDevice({ filters: [] })
    .then(device => {
      assert_unreachable('requestDevice should reject when no device selected');
    })
    .catch(error => {
      assert_equals(error.code, DOMException.NOT_FOUND_ERR);
    })
  );
}, 'requestDevice rejects when no device is chosen');

usb_test(usb => {
  usb.mockDeviceManager.addMockDevice(usb.fakeDevices[0]);
  usb.mockChooserService.setChosenDevice(usb.fakeDevices[0]);
  return callWithKeyDown(() => navigator.usb.requestDevice({ filters: [] })
    .then(device => {
      usb.assertDeviceInfoEquals(device, usb.fakeDevices[0]);
    })
  );
}, 'requestDevice returns the device chosen by the user');

usb_test(usb => {
  const expectedFilters = [
    { vendorId: 1234, classCode: 0xFF, serialNumber: "123ABC" },
    { vendorId: 5678, productId: 0xF00F }
  ];
  usb.mockChooserService.setChosenDevice(null);
  return callWithKeyDown(() => navigator.usb.requestDevice({ filters: expectedFilters })
    .then(device => {
      assert_unreachable('requestDevice should reject because no device selected');
    })
    .catch(error => {
      assert_equals(error.code, DOMException.NOT_FOUND_ERR);
      let actualFilters = usb.mockChooserService.getLastFilters();
      assert_equals(actualFilters.length, expectedFilters.length);
      assert_true(actualFilters[0].has_vendor_id);
      assert_equals(actualFilters[0].vendor_id, expectedFilters[0].vendorId);
      assert_false(actualFilters[0].has_product_id);
      assert_true(actualFilters[0].has_class_code, expectedFilters[0].classCode);
      assert_false(actualFilters[0].has_subclass_code);
      assert_false(actualFilters[0].has_protocol_code);
      assert_equals(actualFilters[0].serial_number, expectedFilters[0].serialNumber);
      assert_true(actualFilters[1].has_vendor_id);
      assert_equals(actualFilters[1].vendor_id, expectedFilters[1].vendorId);
      assert_true(actualFilters[1].has_product_id);
      assert_equals(actualFilters[1].product_id, expectedFilters[1].productId);
      assert_false(actualFilters[1].has_class_code);
      assert_false(actualFilters[1].has_subclass_code);
      assert_false(actualFilters[1].has_protocol_code);
      assert_equals(actualFilters[1].serial_number, null);
    })
  );
}, 'filters are sent correctly');

usb_test(usb => {
  let promise = new Promise((resolve, reject) => {
    navigator.usb.addEventListener('connect', e => {
      assert_true(e instanceof USBConnectionEvent);
      resolve(e.device);
    });
  });

  usb.mockDeviceManager.addMockDevice(usb.fakeDevices[0]);
  return promise.then(device => {
    usb.assertDeviceInfoEquals(device, usb.fakeDevices[0]);
    return device.open().then(() => device.close());
  });
}, 'onconnect event is trigged by adding a device');

usb_test(usb => {
  let promise = new Promise((resolve, reject) => {
    navigator.usb.addEventListener('disconnect', e => {
      assert_true(e instanceof USBConnectionEvent);
      resolve(e.device);
    });
  });

  usb.mockDeviceManager.addMockDevice(usb.fakeDevices[0]);
  usb.mockDeviceManager.removeMockDevice(usb.fakeDevices[0]);
  return promise.then(device => {
    usb.assertDeviceInfoEquals(device, usb.fakeDevices[0]);
    return device.open().then(() => {
      assert_unreachable('should not be able to open a disconnected device');
    }, error => {
      assert_equals(error.code, DOMException.NOT_FOUND_ERR);
    });
  });
}, 'ondisconnect event is triggered by removing a device');
</script>
